Atualização e/ou correção de erros

+++ Desconto.java
@@ -1,13 +1,11 @@
 import javax.swing.JOptionPane;
 
-import javax.swing.JOptionPane; 
-
 class Desconto {
     public static void main(String args[]) {
         JOptionPane jop = new JOptionPane();
 
         double Vcompra = Double.parseDouble(jop.showInputDialog(null, "Valor da compra:"));
-        int Desconto = Double.parseDouble(jop.showInputDialog(null, "Valor do desconto:"));
+        int Desconto = Integer.parseInt(jop.showInputDialog(null, "Valor do desconto:"));
 
         double Dcompra = Vcompra / Desconto;
         double Vfinal = Vcompra - Dcompra;
